requirements.txt
Flask==0.12.1
paho-mqtt==1.2.1
zappa==0.40.0
